imgcompress: Save compressed base image
We want to save the base image so all the data segments are preserved.
For example, in metadata trees there's metadata stored in the data
portions of the leaf nodes. We might eventually want to fuzz that but
for now let's just fuzz the btree metadata.

In the future, it's easy to add more stuff to fuzz by updating the
`CompressedBtrfsImage::metadata`.